backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
dj-database-url>=3.1
python-dotenv>=1.2
requests>=2.32
psycopg[binary]>=3.3
gunicorn>=23.0